Organize package structure and refactor classes in the cart_item package
