+++ src/main/java/com/ecommerce/api/cart/model/Cart.java
@@ -1,6 +1,6 @@
 package com.ecommerce.api.cart.model;
 
-import com.ecommerce.api.cart_item.CartItem;
+import com.ecommerce.api.cart_item.model.CartItem;
 import com.ecommerce.api.user.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
