+++ src/main/java/com/ecommerce/api/cart/response/CartResponse.java
@@ -2,6 +2,7 @@ package com.ecommerce.api.cart.response;
 
 
 import com.ecommerce.api.cart.model.Cart;
+import com.ecommerce.api.cart_item.response.CartItemResponse;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -43,6 +44,6 @@ public class CartResponse implements Serializable {
         private int unitPrice;
         private Long totalPrice;
         private int itemQuantity;
-        private com.ecommerce.api.cart_item.ResponseCartItemDto.ProductDto product;
+        private CartItemResponse.ProductDto product;
     }
 }
