+++ src/main/java/com/ecommerce/api/cart_item/service/CartItemRepository.java
@@ -1,5 +1,6 @@
-package com.ecommerce.api.cart_item;
+package com.ecommerce.api.cart_item.service;
 
+import com.ecommerce.api.cart_item.model.CartItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
