+++ src/main/java/com/ecommerce/api/cart_item/service/CartItemService.java
@@ -1,8 +1,12 @@
-package com.ecommerce.api.cart_item;
+package com.ecommerce.api.cart_item.service;
 
 import com.ecommerce.api.cart.model.Cart;
 import com.ecommerce.api.cart.service.CartRepository;
 import com.ecommerce.api.cart.response.CartResponse;
+import com.ecommerce.api.cart_item.model.CartItem;
+import com.ecommerce.api.cart_item.request.CartItemRequest;
+import com.ecommerce.api.cart_item.response.CartItemResponse;
+import com.ecommerce.api.cart_item.service.CartItemRepository;
 import com.ecommerce.api.product.Product;
 import com.ecommerce.api.product.ProductRepository;
 import com.ecommerce.api.utility.NotFoundException;
@@ -28,9 +32,9 @@ public class CartItemService {
         this.cartRepository = cartRepository;
     }
 
-    public CartResponse createCartItem(RequestCartItemDto requestCartItemDto, Cart cart) {
-        Product product = productRepository.findById(requestCartItemDto.getProductId()).orElseThrow(()->new NotFoundException(requestCartItemDto.getProductId(),Product.class.getName()));
-        CartItem cartItem = this.modelMapper.map(requestCartItemDto, CartItem.class);
+    public CartResponse createCartItem(CartItemRequest cartItemRequest, Cart cart) {
+        Product product = productRepository.findById(cartItemRequest.getProductId()).orElseThrow(()->new NotFoundException(cartItemRequest.getProductId(),Product.class.getName()));
+        CartItem cartItem = this.modelMapper.map(cartItemRequest, CartItem.class);
         cartItem.setUnitPrice(product.getPrice());
         cartItem.setTotalPrice(((long) product.getPrice() * cartItem.getItemQuantity()));
         cartItem.setCart(cart);
@@ -47,7 +51,7 @@ public class CartItemService {
         return modelMapper.map(cartRepository.save(cart), CartResponse.class);
     }
     // Todo fix quantity bug
-    public ResponseCartItemDto updateCartItem(Long Id,int quantity){
+    public CartItemResponse updateCartItem(Long Id, int quantity){
         CartItem cartItem = cartItemRepository.findById(Id).orElseThrow(()->new NotFoundException(Id,CartItem.class.getSimpleName()));
         Cart cart = cartRepository.findById(cartItem.getCart().getCartId()).orElseThrow(()->new NotFoundException(Id,Cart.class.getSimpleName()));
         if(quantity>0){
@@ -60,7 +64,7 @@ public class CartItemService {
         }
         CartItem updatedCartItem = cartItemRepository.save(cartItem);
         cartRepository.save(cart);
-        return this.modelMapper.map(updatedCartItem, ResponseCartItemDto.class);
+        return this.modelMapper.map(updatedCartItem, CartItemResponse.class);
     }
 
     public void deleteCartItem(Long Id){
