+++ src/main/java/com/ecommerce/api/product/Product.java
@@ -1,6 +1,6 @@
 package com.ecommerce.api.product;
 
-import com.ecommerce.api.cart_item.CartItem;
+import com.ecommerce.api.cart_item.model.CartItem;
 import com.ecommerce.api.order_item.OrderItem;
 import com.ecommerce.api.utility.EnumConstants;
 import jakarta.persistence.*;
